road_detection_api/requirements.txt
matplotlib==3.8.2
tensorflow==2.15.0
scikit-learn==1.3.2
scikit-image==0.22.0
keras==2.15.0
python-dotenv==1.0.0
fastapi
uvicorn[standard]
opencv-python
rich==13.7.0